refactor: create error handler module
