+++ src/error-handler/hooks/useErrorHandler.tsx
@@ -1,0 +1,9 @@
+import React, { Suspense } from 'react';
+
+export default function useErrorHandler(callBack) {
+  try {
+    await callBack();
+  } catch (error) {
+    if ()
+  }
+}
